+++ easy-ui-react/src/ProductLayout/ProductLayoutSidebar.tsx
@@ -1,6 +1,6 @@
 import React, { ReactNode } from "react";
 
-import styles from "./ProductLayout.module.scss";
+import styles from "./ProductLayoutSidebar.module.scss";
 
 export type ProductLayoutSidebarProps = {
   children: ReactNode;
@@ -9,7 +9,11 @@ export type ProductLayoutSidebarProps = {
 export function ProductLayoutSidebar(props: ProductLayoutSidebarProps) {
   const { children } = props;
   return (
-    <div role="region" aria-label="Sidebar" className={styles.sidebar}>
+    <div
+      role="region"
+      aria-label="Sidebar"
+      className={styles.ProductLayoutSidebar}
+    >
       {children}
     </div>
   );
