+++ easy-ui-react/src/ProductLayout/ProductLayoutTabbedContent.module.scss
@@ -1,0 +1,11 @@
+@use "../styles/common" as *;
+@use "../styles/unstyled";
+
+.ProductLayoutTabbedContent {
+  display: flex;
+  flex-direction: column;
+  background: white;
+  border-radius: 4px;
+  padding: 24px;
+  padding-top: 12px;
+}
